remove deprecated vars
